Switch from pr flag to prs command
add an output file flag and support for multiple labels

also add a issues command in preparation to add
support for issues, currently just returns prs data
