+++ .gitignore
@@ -30,3 +30,5 @@ gha*
 
 # Added by goreleaser init:
 dist/
+
+.idea
