+++ cmd/root.go
@@ -21,8 +21,8 @@ var thisweek bool
 var today bool
 
 var repo string
-var label string
-var prs bool
+var labels []string
+var output string
 
 // rootCmd represents the base command when called without any subcommands
 var rootCmd = &cobra.Command{
@@ -30,10 +30,6 @@ var rootCmd = &cobra.Command{
 	Short: "Get your Github activity",
 	Long:  `Get PRs, reviews, and issues created during a specific time interval.`,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		if prs {
-			return github.GetPRData(domain, token, repo, label)
-		}
-
 		if lastweek == true {
 			startdate, enddate = utils.GetLastWeekDates()
 		} else if thisweek == true {
@@ -57,14 +53,11 @@ func Execute() {
 
 func init() {
 	rootCmd.PersistentFlags().StringVarP(&domain, "domain", "d", "github.com", "Github domain")
-	rootCmd.PersistentFlags().StringVarP(&startdate, "start", "s", utils.GetDefaultStartDate(), "Collect activities starting on this date")
-	rootCmd.PersistentFlags().StringVarP(&enddate, "end", "e", utils.GetDefaultEndDate(), "Collect activities up to this date")
-	rootCmd.PersistentFlags().BoolVarP(&lastweek, "last-week", "l", false, "Collect activities for last week (last week Monday to last week Friday")
-	rootCmd.PersistentFlags().BoolVarP(&thisweek, "this-week", "w", false, "Collect activities for this week (Monday to Friday")
-	rootCmd.PersistentFlags().BoolVarP(&today, "today", "n", false, "Collect activities for today")
+	rootCmd.Flags().StringVarP(&startdate, "start", "s", utils.GetDefaultStartDate(), "Collect activities starting on this date")
+	rootCmd.Flags().StringVarP(&enddate, "end", "e", utils.GetDefaultEndDate(), "Collect activities up to this date")
+	rootCmd.Flags().BoolVarP(&lastweek, "last-week", "l", false, "Collect activities for last week (last week Monday to last week Friday)")
+	rootCmd.Flags().BoolVarP(&thisweek, "this-week", "w", false, "Collect activities for this week (Monday to Friday)")
+	rootCmd.Flags().BoolVarP(&today, "today", "n", false, "Collect activities for today")
 	rootCmd.PersistentFlags().StringVarP(&username, "user", "u", utils.GetCurrentUsername(), "Username")
 	rootCmd.PersistentFlags().StringVarP(&token, "token", "t", "", "Github Personal Access Token (default `$GITHUB_TOKEN`)")
-	rootCmd.PersistentFlags().BoolVarP(&prs, "prs", "p", false, "Return PR data")
-	rootCmd.PersistentFlags().StringVarP(&repo, "repo", "r", "i-am-bee/beeai-framework", "Github org/repo")
-	rootCmd.PersistentFlags().StringVar(&label, "label", "python", "Issue/PR label")
 }
